Switch to Google Maps
And it all starts working.
