+++ vs-ui/package.json
@@ -13,12 +13,11 @@
     "@turf/bbox": "^6.3.0",
     "fhir-kit-client": "^1.6.5",
     "formik": "^2.2.6",
+    "google-map-react": "^2.1.9",
     "history": "^5.0.0",
-    "mapbox-gl": "^1.12.0",
     "node-sass": "^5.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-mapbox-gl": "^5.1.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "typescript": "^4.1.2",
@@ -50,10 +49,9 @@
   "devDependencies": {
     "@testing-library/dom": "^7.30.3",
     "@types/fhir-kit-client": "^1.1.1",
-    "@types/geojson": "^7946.0.7",
+    "@types/google-map-react": "^2.1.0",
     "@types/history": "^4.7.8",
     "@types/jest": "^26.0.15",
-    "@types/mapbox-gl": "^1.12.8",
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
