+++ vs-ui/src/components/VSMap/utils.tsx
@@ -28,7 +28,8 @@ export function notEmpty<TValue>(
   value: TValue | null | undefined
 ): value is TValue {
   if (value === null || value === undefined) return false;
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  // noinspection JSUnusedLocalSymbols
+  // eslint-disable-next-line
   const testDummy: TValue = value;
   return true;
 }
